requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pdfplumber
python-dotenv
openai
pydantic-settings
python-multipart
